Rename all instances of subNavigation to sideNavigation
