+++ site/src/App/routes/foundations/iconography/iconography.tsx
@@ -17,7 +17,7 @@ const Iconography = () => (
       Iconography
     </Heading>
 
-    <DocNavigationBar title="Subnavigation">
+    <DocNavigationBar title="Sidenavigation">
       <DocNavigationItem href="/foundations/iconography">
         Details
       </DocNavigationItem>
